requirements.txt
flask
tensorflow-cpu
pillow
numpy
jinja2
gunicorn
itsdangerous
MarkupSafe
Werkzeug
